DROID-2493 move methods
